more v33 renaming and removal of dead code
