+++ renderlib/glsl/CMakeLists.txt
@@ -10,12 +10,8 @@ target_sources(renderlib PRIVATE
 	"${CMAKE_CURRENT_SOURCE_DIR}/GLPTVolumeShader.h"	
 	"${CMAKE_CURRENT_SOURCE_DIR}/GLToneMapShader.cpp"
 	"${CMAKE_CURRENT_SOURCE_DIR}/GLToneMapShader.h"	
-	"${CMAKE_CURRENT_SOURCE_DIR}/V330GLFlatShader2D.cpp"
-	"${CMAKE_CURRENT_SOURCE_DIR}/V330GLFlatShader2D.h"	
-	"${CMAKE_CURRENT_SOURCE_DIR}/V330GLImageShader2D.cpp"
-	"${CMAKE_CURRENT_SOURCE_DIR}/V330GLImageShader2D.h"	
-	"${CMAKE_CURRENT_SOURCE_DIR}/V330GLImageShader2DnoLut.cpp"
-	"${CMAKE_CURRENT_SOURCE_DIR}/V330GLImageShader2DnoLut.h"	
-	"${CMAKE_CURRENT_SOURCE_DIR}/V330GLLineShader2D.cpp"
-	"${CMAKE_CURRENT_SOURCE_DIR}/V330GLLineShader2D.h"	
+	"${CMAKE_CURRENT_SOURCE_DIR}/GLFlatShader2D.cpp"
+	"${CMAKE_CURRENT_SOURCE_DIR}/GLFlatShader2D.h"	
+	"${CMAKE_CURRENT_SOURCE_DIR}/GLImageShader2DnoLut.cpp"
+	"${CMAKE_CURRENT_SOURCE_DIR}/GLImageShader2DnoLut.h"	
 )
