+++ renderlib/glsl/GLImageShader2DnoLut.cpp
@@ -1,4 +1,4 @@
-#include "V330GLImageShader2D.h"
+#include "GLImageShader2DnoLut.h"
 
 #include "Logging.h"
 
@@ -8,7 +8,7 @@
 #include <iostream>
 #include <sstream>
 
-GLImageShader2D::GLImageShader2D()
+GLImageShader2DnoLut::GLImageShader2DnoLut()
   : GLShaderProgram()
   , m_vshader()
   , m_fshader()
@@ -16,9 +16,6 @@ GLImageShader2D::GLImageShader2D()
   , m_attr_texcoords()
   , m_uniform_mvp()
   , m_uniform_texture()
-  , m_uniform_lut()
-  , m_uniform_min()
-  , m_uniform_max()
 {
   m_vshader = new GLShader(GL_VERTEX_SHADER);
 
@@ -39,17 +36,13 @@ GLImageShader2D::GLImageShader2D()
                                "}\n");
 
   if (!m_vshader->isCompiled()) {
-    LOG_ERROR << "GLImageShader2D: Failed to compile vertex shader\n" << m_vshader->log();
+    LOG_ERROR << "GLImageShader2DnoLut: Failed to compile vertex shader\n" << m_vshader->log();
   }
 
   m_fshader = new GLShader(GL_FRAGMENT_SHADER);
   m_fshader->compileSourceCode("#version 400 core\n"
                                "\n"
                                "uniform sampler2D tex;\n"
-                               "uniform sampler1DArray lut;\n"
-                               "uniform vec3 texmin;\n"
-                               "uniform vec3 texmax;\n"
-                               "uniform vec3 correction;\n"
                                "\n"
                                "in VertexData\n"
                                "{\n"
@@ -62,12 +55,11 @@ GLImageShader2D::GLImageShader2D()
                                "  vec2 flipped_texcoord = vec2(inData.f_texcoord.x, 1.0 - inData.f_texcoord.y);\n"
                                "  vec4 texval = texture(tex, flipped_texcoord);\n"
                                "\n"
-                               "  outputColour = texture(lut, vec2(((((texval[0] * correction[0]) - texmin[0]) / "
-                               "(texmax[0] - texmin[0]))), 0.0));\n"
+                               "  outputColour = texval;\n"
                                "}\n");
 
   if (!m_fshader->isCompiled()) {
-    LOG_ERROR << "GLImageShader2D: Failed to compile fragment shader\n" << m_fshader->log();
+    LOG_ERROR << "GLImageShader2DnoLut: Failed to compile fragment shader\n" << m_fshader->log();
   }
 
   addShader(m_vshader);
@@ -75,133 +67,97 @@ GLImageShader2D::GLImageShader2D()
   link();
 
   if (!isLinked()) {
-    LOG_ERROR << "GLImageShader2D: Failed to link shader program\n" << log();
+    LOG_ERROR << "GLImageShader2DnoLut: Failed to link shader program\n" << log();
   }
 
   m_attr_coords = attributeLocation("coord2d");
   if (m_attr_coords == -1)
-    LOG_ERROR << "GLImageShader2D: Failed to bind coordinates";
+    LOG_ERROR << "GLImageShader2DnoLut: Failed to bind coordinates";
 
   m_attr_texcoords = attributeLocation("texcoord");
   if (m_attr_texcoords == -1)
-    LOG_ERROR << "GLImageShader2D: Failed to bind texture coordinates";
+    LOG_ERROR << "GLImageShader2DnoLut: Failed to bind texture coordinates";
 
   m_uniform_mvp = uniformLocation("mvp");
   if (m_uniform_mvp == -1)
-    LOG_ERROR << "GLImageShader2D: Failed to bind transform";
+    LOG_ERROR << "GLImageShader2DnoLut: Failed to bind transform";
 
   m_uniform_texture = uniformLocation("tex");
   if (m_uniform_texture == -1)
-    LOG_ERROR << "GLImageShader2D: Failed to bind texture uniform ";
-
-  m_uniform_lut = uniformLocation("lut");
-  if (m_uniform_lut == -1)
-    LOG_ERROR << "GLImageShader2D: Failed to bind lut uniform ";
-
-  m_uniform_min = uniformLocation("texmin");
-  if (m_uniform_min == -1)
-    LOG_ERROR << "GLImageShader2D: Failed to bind min uniform ";
-
-  m_uniform_max = uniformLocation("texmax");
-  if (m_uniform_max == -1)
-    LOG_ERROR << "GLImageShader2D: Failed to bind max uniform ";
-
-  m_uniform_corr = uniformLocation("correction");
-  if (m_uniform_corr == -1)
-    LOG_ERROR << "GLImageShader2D: Failed to bind correction uniform ";
+    LOG_ERROR << "GLImageShader2DnoLut: Failed to bind texture uniform ";
 }
 
-GLImageShader2D::~GLImageShader2D() {}
+GLImageShader2DnoLut::~GLImageShader2DnoLut() {}
 
 void
-GLImageShader2D::enableCoords()
+GLImageShader2DnoLut::enableCoords()
 {
   enableAttributeArray(m_attr_coords);
+  check_gl("enable attribute array: coords");
 }
 
 void
-GLImageShader2D::disableCoords()
+GLImageShader2DnoLut::disableCoords()
 {
   disableAttributeArray(m_attr_coords);
+  check_gl("disable attribute array: coords");
 }
 
 void
-GLImageShader2D::setCoords(const GLfloat* offset, int tupleSize, int stride)
+GLImageShader2DnoLut::setCoords(const GLfloat* offset, int tupleSize, int stride)
 {
   setAttributeArray(m_attr_coords, offset, tupleSize, stride);
+  check_gl("set attr coords pointer");
 }
 
 void
-GLImageShader2D::setCoords(GLuint coords, const GLfloat* offset, int tupleSize, int stride)
+GLImageShader2DnoLut::setCoords(GLuint coords, const GLfloat* offset, int tupleSize, int stride)
 {
   glBindBuffer(GL_ARRAY_BUFFER, coords);
   setCoords(offset, tupleSize, stride);
   glBindBuffer(GL_ARRAY_BUFFER, 0);
+  check_gl("set attr coords buffer");
 }
 
 void
-GLImageShader2D::enableTexCoords()
+GLImageShader2DnoLut::enableTexCoords()
 {
   enableAttributeArray(m_attr_texcoords);
+  check_gl("enable attribute array texcoords");
 }
 
 void
-GLImageShader2D::disableTexCoords()
+GLImageShader2DnoLut::disableTexCoords()
 {
   disableAttributeArray(m_attr_texcoords);
+  check_gl("disable attribute array texcoords");
 }
 
 void
-GLImageShader2D::setTexCoords(const GLfloat* offset, int tupleSize, int stride)
+GLImageShader2DnoLut::setTexCoords(const GLfloat* offset, int tupleSize, int stride)
 {
   setAttributeArray(m_attr_texcoords, offset, tupleSize, stride);
+  check_gl("set attr texcoords ptr");
 }
 
 void
-GLImageShader2D::setTexCoords(GLuint texcoords, const GLfloat* offset, int tupleSize, int stride)
+GLImageShader2DnoLut::setTexCoords(GLuint texcoords, const GLfloat* offset, int tupleSize, int stride)
 {
   glBindBuffer(GL_ARRAY_BUFFER, texcoords);
   setTexCoords(offset, tupleSize, stride);
   glBindBuffer(GL_ARRAY_BUFFER, 0);
+  check_gl("set attr texcoords buffer");
 }
 
 void
-GLImageShader2D::setTexture(int texunit)
+GLImageShader2DnoLut::setTexture(int texunit)
 {
   glUniform1i(m_uniform_texture, texunit);
   check_gl("Set image texture");
 }
 
 void
-GLImageShader2D::setMin(const glm::vec3& min)
-{
-  glUniform3fv(m_uniform_min, 1, glm::value_ptr(min));
-  check_gl("Set min range");
-}
-
-void
-GLImageShader2D::setMax(const glm::vec3& max)
-{
-  glUniform3fv(m_uniform_max, 1, glm::value_ptr(max));
-  check_gl("Set max range");
-}
-
-void
-GLImageShader2D::setCorrection(const glm::vec3& corr)
-{
-  glUniform3fv(m_uniform_corr, 1, glm::value_ptr(corr));
-  check_gl("Set correction multiplier");
-}
-
-void
-GLImageShader2D::setLUT(int texunit)
-{
-  glUniform1i(m_uniform_lut, texunit);
-  check_gl("Set LUT texture");
-}
-
-void
-GLImageShader2D::setModelViewProjection(const glm::mat4& mvp)
+GLImageShader2DnoLut::setModelViewProjection(const glm::mat4& mvp)
 {
   glUniformMatrix4fv(m_uniform_mvp, 1, GL_FALSE, glm::value_ptr(mvp));
   check_gl("Set image2d uniform mvp");
